+++ index.html
@@ -13,7 +13,7 @@
     <script
       src="https://kit.fontawesome.com/5e248e42c9.js"
       crossorigin="anonymous"></script>
-      <script src="js/login/test.js" type="module"></script>
+      <script src="js/login/main.js" type="module"></script>
   </head>
   <body>
     <header class="fixed-top">
@@ -41,7 +41,7 @@
                 <a class="nav-link" href="/feed/">Feed</a>
               </li>
               <li class="nav-item"></li>
-                <a class="nav-link" href="/index.html">Log Out</a>
+                <button type="submit" id="logout">Log Out</button>
               </li>
             </ul>
