+++ js/load.js
@@ -1,0 +1,3 @@
+export function load(key) {
+    return JSON.parse(localStorage.getItem(key));
+}
