+++ js/save.js
@@ -1,0 +1,3 @@
+export function save(key, value) {
+    localStorage.setItem(key, JSON.stringify(value));
+}
