+++ js/url.js
@@ -1,0 +1,7 @@
+export const apiBase = 'https://v2.api.noroff.dev';
+export const apiAuth = '/auth';
+export const apiReg = '/register';
+export const apiLogin = '/login';
+export const apiPosts = '/social/posts';
+export const apiAuthor = '/?_author=true'
+export const apiKey = '4f3ee179-18e9-46f1-b721-844731846040';
