--- js_not_in_use/test.html
@@ -1,110 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Login/Register</title>
-    <style>
-      body {
-        font-family: Arial, sans-serif;
-        margin: 0;
-        padding: 0;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        height: 100vh;
-        background-color: #f0f0f0;
-      }
-      .auth-container {
-        background: #fff;
-        padding: 20px;
-        border-radius: 8px;
-        box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-        width: 100%;
-        max-width: 400px;
-      }
-      .auth-container h1 {
-        margin: 0 0 20px;
-        font-size: 24px;
-        text-align: center;
-      }
-      .auth-container .form-group {
-        margin-bottom: 15px;
-      }
-      .auth-container label {
-        display: block;
-        font-size: 14px;
-        margin-bottom: 5px;
-      }
-      .auth-container input {
-        width: 100%;
-        padding: 10px;
-        border: 1px solid #ccc;
-        border-radius: 4px;
-        font-size: 14px;
-      }
-      .auth-container .button-group {
-        display: flex;
-        justify-content: space-between;
-        gap: 10px;
-      }
-      .auth-container button {
-        flex: 1;
-        padding: 10px;
-        font-size: 14px;
-        border: none;
-        border-radius: 4px;
-        cursor: pointer;
-      }
-      .auth-container button.login {
-        background-color: #007bff;
-        color: #fff;
-      }
-      .auth-container button.register {
-        background-color: #28a745;
-        color: #fff;
-      }
-    </style>
-  </head>
-  <body>
-    <div class="auth-container">
-      <h1>Login/Register</h1>
-      <form name="auth">
-        <div class="form-group">
-          <label for="name">Name (for Register)</label>
-          <input
-            type="text"
-            id="name"
-            name="name"
-            placeholder="Enter your name" />
-        </div>
-        <div class="form-group">
-          <label for="email">Email</label>
-          <input
-            type="email"
-            id="email"
-            name="email"
-            placeholder="Enter your email"
-            required />
-        </div>
-        <div class="form-group">
-          <label for="password">Password</label>
-          <input
-            type="password"
-            id="password"
-            name="password"
-            placeholder="Enter your password"
-            required />
-        </div>
-        <div class="button-group">
-          <button type="submit" class="login" data-auth="login">Login</button>
-          <button type="submit" class="register" data-auth="register">
-            Register
-          </button>
-        </div>
-      </form>
-    </div>
-
-    <script src="test.js" type="module"></script>
-  </body>
-</html>
