--- js_not_in_use/userToLogin.js
@@ -1,5 +1,0 @@
-export const userToLogin = {
-    name: "testsdweghname", 
-    email: "testfirstlast@noroff.no", 
-    password: "testeUzI1NiIsInR5cCI",
-}
