--- js_not_in_use/userToRegister.js
@@ -1,5 +1,0 @@
- export const userToRegister = {
-    name: "testsdweghname", 
-    email: "testfirstlast@noroff.no", 
-    password: "testeUzI1NiIsInR5cCI"
-}
